Add unit tests for nashorn java script engine
